requirements.txt
google-cloud-vision
python-dotenv
vertexai
tqdm
requests
beautifulsoup4
selenium
webdriver-manager
Pillow

